+++ crates/egui_demo_lib/src/demo/mod.rs
@@ -12,6 +12,7 @@ pub mod dancing_strings;
 pub mod demo_app_windows;
 pub mod drag_and_drop;
 pub mod font_book;
+pub mod highlighting;
 pub mod layout_test;
 pub mod misc_demo_window;
 pub mod multi_touch;